requirements.txt
Pillow==10.0.0
pytesseract==0.3.10
pyautogui==0.9.54
keyboard==0.13.5
pyperclip==1.8.2
pyinstaller==6.3.0
